Changed inclusion of body parameter in signature
Hi,

I changed the inclusion of body parameters in signature to be optional. I just added a boolean to decide if parameters have to be signed.

For example, I use Java CXF framework to implement a OAuth 1.0a server in my backend and PHP + Guzzle to query this backend and the CXF framework don't sign GET and POST parameters. 

So I think it may be useful to let end user choose if it have to be signed. 

Thank you